feat: components destructure work
