+++ components/SignUp/SignUpControl.tsx
@@ -1,27 +1,46 @@
-import React, { ChangeEvent, useCallback, useState } from 'react';
-import {
-  SignUpSection,
-  MainHeadText,
-  SubHeadText,
-  SignUpBtn,
-  StyledInputBox,
-  StyledInput,
-  Checkbox,
-  SignUpCheckBox,
-  SignUpCheckBoxLayout,
-  StyledLabel,
-  StyledLabelAll,
-  Line,
-  TextErrorParagraph,
-  StyledLabelEssential,
-  TextParagraphSns,
-  SnsLine,
-  TextParagraphInfo,
-  TextParagrapValidate,
-} from '@/styles/signUp';
-import { useMutation } from 'react-query';
-import { addUser } from '@/api/user';
-import { NextRouter, useRouter } from 'next/router';
-import MobileBottomNav from '../common/MobileBottomNav';
-import SocialLogin from '../SignIn/SocialLogIn';
-import { SecretEye } from '../common/SVG';
+import React from 'react';
+import { SignUpBtn } from '@/styles/signUp';
+
+interface Props {
+  signUpState: {
+    email: string;
+    password: string;
+    passwordConfirm: string;
+    nickname: string;
+  };
+}
+
+const SignUpcontrol: React.FC<Props> = ({ signUpState }) => {
+  return (
+    <SignUpBtn
+      style={{
+        margin: '12px 0 20px 0',
+        backgroundColor:
+          signUpState.email &&
+          signUpState.password &&
+          signUpState.passwordConfirm &&
+          signUpState.nickname
+            ? '#0078FF'
+            : '#B3B3B3',
+        borderColor:
+          signUpState.email &&
+          signUpState.password &&
+          signUpState.passwordConfirm &&
+          signUpState.nickname
+            ? '#0078FF'
+            : '#B3B3B3',
+      }}
+      disabled={
+        !signUpState.email &&
+        !signUpState.password &&
+        !signUpState.passwordConfirm &&
+        !signUpState.nickname
+      }
+      type="submit"
+    >
+      회원가입하기
+    </SignUpBtn>
+  );
+};
+
+export default React.memo(SignUpcontrol);
